refactor: rename permission rules to rules
